requirements.txt
pymongo
gdeltdoc
newspaper3k
openai
streamlit
pandas
